requirements/defaults.txt
Django==1.8.6
anyjson==0.3.3
beautifulsoup4==4.4.0
billiard==3.3.0.20
celery==3.1.18
coverage==3.7.1
dj-database-url==0.3.0
dj-static==0.0.6
djangorestframework==3.2.3
gunicorn==19.3.0
kombu==3.0.26
langdetect==1.0.5
paramiko==1.15.2
psycopg2==2.6
requests==2.6.2
six==1.9.0
static3==0.6.1
